backend/requirements.txt
# requirements_updated.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.34
cx-oracle==8.3.0
oracledb==2.0.0  #
apscheduler==3.10.4
aiosmtplib==3.0.1
python-dotenv==1.0.0
pydantic>=2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator>=2.1.1